archivage/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-auditlog==3.0.0
pillow==10.3.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
